app/scheduler/pyproject.toml
[tool.poetry]
name = "resaas-scheduler"
version = "0.1.0"
description = "Job scheduler for resaas"
authors = ["dorranh <dorran.howell@tweag.io>"]

[tool.poetry.dependencies]
python = "^3.8"
resaas-common = { path = "../../lib/common", develop = true }
Flask = "^1.1.2"
apache-libcloud = "^3.3.0"
marshmallow = "^3.10.0"
SQLAlchemy = "^1.3.22"
psycopg2-binary = "^2.8.6"
paramiko = "^2.7.2"
Flask-SQLAlchemy = "^2.4.4"
flask-marshmallow = "^0.14.0"
marshmallow-sqlalchemy = "^0.24.1"
marshmallow-enum = "^1.5.1"
PyYAML = "^5.3.1"
shortuuid = "^1.0.1"
celery = {extras = ["gevent", "redis"], version = "^5.0.5"}
Flask-Cors = "^3.0.10"
grpcio = "^1.35.0"
resaas-grpc = { path = "../../lib/grpc", develop = true }

[tool.poetry.dev-dependencies]
pytest = "^6.2.1"
black = "^20.8b1"
flake8 = "^3.8.4"
pylint = "^2.6.0"
mypy = "^0.790"
pytest-dotenv = "^0.5.2"
pytest-cov = "^2.10.1"
isort = "^5.7.0"

[tool.black]
line-length = 99
target-version = ["py38"]
[tool.pytest.ini_options]
minversion = "6.0"
env_override_existing_values = 1
env_files = [".env"]

[tool.coverage.run]
omit = ["*test*"]
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

